requirements.txt
openai
pydantic
python-dotenv
pyyaml
websockets